cockpit: Don't clobber an existing bots checkout
Commit fded3bd579e was a thinko -- for our CI we *don't* want our test to
clobber a pre-existing bots/ checkout, as we often use this to run tests
against an updated image or to validate a changes to the bots project.

On developer machines, bots may also be a symlink to an actual bots
directory in development, so don't clobber that.

Ported from https://github.com/cockpit-project/starter-kit/pull/232