webpack: Use relative resolve path for npm 7 compatibility
npm 7 changed how it resolves dependencies, and cockpit fails to
build with lots of unresolved peer dependencies of PatternFly.

With an absolute path, `resolve.modules` will only look in that
directory; the default is a relative path "node_modules" that just
works [1]. We need to keep the `$SRCDIR` support, but convert the path
to a relative one to keep the old recursive search behaviour.

This was originally applied as commit 8c68b1ed855, and reverted in
9083c54a5d9 as it broke the release due to [2]. That has a workaround
now.

[1] https://webpack.js.org/configuration/resolve/#resolvemodules
[2] https://github.com/webpack/webpack/issues/14481

See cockpit commit a117600dfffe8311ec6286cf060c80cd35a8a84c