Drop the Vagrant bits
The current bits for Vagrant are broken, and it seems there is not much
interest from the current development base to actively maintain and use
them. Hence, to avoid more bit-rotting, simply drop all the Vagrant
bits, and references to them where not useful for something else.

Remove also a couple of helper scripts that use the Vagrant bits, as
they clearly cannot work.