Initialize the tito build system
Add the basic configuration to use tito, with a custom tagger that
increments the current version by default.

Since tito has its own way of building upstream tarballs which is not
easily compatible with the current buildsystem, then provide a custom
builder that always errors out. There are convenience targets for
creating a tarball ("dist"), building a source RPM ("srpm"), and
building the binary RPM packages ("rpm"), so there is way less need for
the support of tito.