Move some data files under 'data' subdirectory
Move some metadata files (not strictly the cockpit plugin itself) to a
'data' subdirectory. They are not changed often, so there is no need to
have them around e.g. in the top-level directory of the sources.

Move also the icons under that subdirectory; since they are used by the
subscription-manager dist tarball without being included in it (sigh),
then we need to make sure they end up in the old path, i.e. where the
build system of subscription-manager expects them.