tests: allow TEST_SCENARIO=system
The test suite already uses TEST_SCENARIO to specify which branch of
subscription-manager checkout & test, defaulting to "main". While this
works nicely, it unconditionally tests subscription-manager from
upstream sources.

Introduce a new special value for TEST_SCENARIO, i.e. "system": this
does not represent any special branch, but simply avoids installing
subscription-manager from sources, and relies on it being already
installed in the test image. Document the existing behaviour of
TEST_SCENARIO, and this new addition, in our HACKING file.

The actual implementation is simple: all is needed for a "system"
scenario is to not pass to image-customize the tarballs & scripts of
subscription-manager.