tests: bump image-customize RAM to 2048
The default amount of RAM for the VM used by image-customize is ~1100MB,
which seems to not be enough for webpack to fully run (as invoked during
the RPM %build phase) on RHEL 8.

While a possible long term solution could be to reduce/update the
nodejs modules, for now simply bump the RAM used by image-customize to
allow webpack to complete.