test: simplify machine_python()
Compose the python3 command to execute, arguments included, as list to
pass to machine.execute(): that function already properly takes care
about quoting for execution on the machine, so there is no need to do it
manually ourselves. Also generalize the handling of parameters for the
Python script, so there is no limit of 1.

shell_escape() is dropped, as was used only in machine_python().