test: improve calls to Machine.execute()
This function already properly takes care about quoting for execution
on the machine, so there is no need to do it manually ourselves in all
cases. Hence, switch cases of plain commands with arguments (with no
need for shell) from a command string to a list of the command and its
arguments.

There should be no behaviour changes.