ci: add black & flake8 check using GitHub Actions
Check the Python sources with black and flake8 on PRs using GitHub
Actions. The job will install flake8 as available in the latest stable
Fedora version, and black from pip with the precise version we want.

Use an official PSF action to run black, and a marketplace action to
annotate the files with the result of the flake8 run.