insights: improve handling of registration/unregistration errors
Add an helper function to improve the way the catch() handler of a
cockpit.spawn() produces the string representation of the error;
use it in the existing catch_error() function, used as catch() handler
when running `insights-client --register/--unregister`.

This makes the floating error alert show what was the issue with
insights-client, however only in the case of attempts to Insights
registration during the system registration (i.e. not when trying to
register to Insights from its own dedicated dialog).