insights: properly show error message in Insights dialog
The Insights dialog is a cockpit.Dialog component, which has a
cockpit.DialogFooter element to handle its footer and that includes
actions. The default handler of those actions expects a promise,
and its catch() handlers gets the error message from it. The problem is
that the promise in this case is what cockpit.spawn() returns, and thus
the error message is lost:
- the second parameter of the cockpit.spawn() error handler contains
  the complete output
- the cockpit.DialogFooter catch() handler uses only the first
  parameter, which is the exception object

Hence, catch the cockpit.spawn() promise ourselves, creating a custom
promise for cockpit.DialogFooter:
- in case of a success, the new promise is resolved
- in case of failure, we use the existing spawn_error_to_string() helper
  function to properly format the cockpit.spawn() failure, creating a
  minimal fake exception object that cockpit.DialogFooter will use to
  fill the inline error <Alert>

The only problem is that, due to the current version of the cockpit
components in use (253), cockpit.DialogFooter can only fill the title of
its <Alert> elements, and I have found no way to make multiline text
properly formatted. This will be doable once we upgrade to at least
cockpit 267 for the components, as cockpit.DialogFooter since that
version supports filling the details of <Alert> objects.