Switch away from 'throw'
'throw' will raise a synchronous exception which bypasses all the error
handling based on defer/promises, triggering an "Ooops!" error window of
cockpit, and leaving an in-progress register dialog.

Instead, use the standard defer/promise error handling done in other
parts of the registerSystem() function already, so the error message
will be properly shown as inline error in the register dialog.