tito: remove autogenerated spec file before tagging
tito looks for the .spec file first as ".spec", then as ".spec.tmpl";
the former is autogenerated by the build system from the latter, using
the version of the package. In case the tree is built, and thus contains
an autogenerated .spec file, tito will pick it, modifying it, and trying
to commit it, which will eventually fail as it is an ignored file for
git.

As a way to unconfuse tito, remove the autogenerated .spec file before
building the low-level VersionTagger object, so tito will find the right
file.