spec: suggest insights-client on RHEL
The Cockpit plugin can register the system to Red Hat Insights using
insights-client (which can be installed if missing); hence it makes
sense to add some package relationship for it.

Considering that insights-client exists only on RHEL, limit the
insights-client relationship to RHEL only.

Also, insights-client is in the AppStream repository, while the Cockpit
plugin is in BaseOS; because of that, there cannot be neither an hard
Requires nor a Recommends, hence the only way is a Suggests.