2169976: insights: fix spawn_error_to_string in more cases
In case the error has a "message" field, let cockpit.message() convert
it to string, as it knows how to do that properly.

Also, in case we need to use the string representation of the error as
error string, force it to be a string, instead of returning an object.

Fixes/followups commit fc725f7dd1d83fcc8bc2a8346e4bbd19d27a0b5b and
commit da3457530dc15172f43c97df01122df35272ef5d.