Update cockpit test lib and webpack build
The most recent Chromium 113 does not work any more with sizzle, tests
were hanging as soon as they switched to a frame. The latest cockpit
testlib applied a workaround [1], pull that in.

Due to the conversion from CommonJS to ESM, this requires updating
Cockpit's pkg/lib as well. Update the webpack NPM modules and
configuration to the most recent starter-kit version (commit 0aaccb9ea1,
right before the conversion to esbuild):

 * Use ESM `import` instead of `require()`.

 * This actually *uses* terser and CSS minifier to reduce the webpack
   size, instead of just importing it in vain.

 * Drop the unused and unsupportable srcdir/builddir variables for
   building in a separate directory.

 * Don't compress CSS in devel mode.

 * Drop the obsolete md4 hack.

[1] https://github.com/cockpit-project/cockpit/pull/18812