Fix obsolete cockpit defer API
done/fail/always is the long-obsolete `cockpit.defer()` jQuery promise
API, which will be dropped eventually. Replace them with the standard
`Promise` API.

This requires a little teak in safeDBusCall() -- while `.fail()` also
returns the rejection, `Promise.catch()` handles it and returns nothing.

In `InstalledProducts.handleAutoAttach()`, both handlers in were doing
exactly the same thing, so use finally() instead. Also use an arrow
instead of a function to avoid the `self = this` dance.