test: break insights-client in a different way
The sequence of events in the test is the following:
1) register with insights-client, a new machine-id is generated
2) the test moves away the machine-id file
3) a new data upload is triggered by starting the systemd service
4) insights-client eagerly generates a new machine-id file to check
   whether it is registered even if it is not the case (that's another
   issue, much harder to fix)
5) the newly generated machine-id has nothing associated to it
6) a proper warning is shown as expected
7) the original machine-id is restored
8) everything else continues as usual

This worked fine until recent times, as the new machine-id always had
a new UUID. After recent changes in insights-core [1], now the default
machine-id generated by insights-client when the system is registered
with subscription-manager is the subscription-manager UUID. Thus, the
machine-id file generated during step (4) is now always the same, which
then allows insights-client to query for its own details as if nothing
happened.

To break insights-client again, the changes are in the following step
of the above sequence:
4) a new random UUID is written as machine-id file

This new "wrong" machine-id file makes insights-client think that the
status is not correct, and thus the test can keep going as expected.

[1] https://github.com/RedHatInsights/insights-core/pull/4057