feat: switch from .last-upload.results to .lastupload
The file /etc/insights-client/.last-upload.results is watched to
determine when was the last upload done. A couple of important notes
about it:
1) even though it is read (and parsed), its content is actually never
   used currently
2) the file is written by insights-client (better: insights-core) only
   when "legacy_upload" is true (which is the default as of this
   writing)

While (1) is not much an issue, (2) means that the displaying of the
last upload (timestamp) to Insights is not available when switching
insights-client to non-legacy-upload.

To avoid both the issues, switch from .last-upload.results to
/etc/insights-client/.lastupload:
- .lastupload is always written, no matter the mode, and its
  modification timestamp is indeed accurate
- stop reading the file at all, saving a bit of resources

There should be no behaviour changes.