test: setup the subscription repositories
Ensure that subscription-manager is configured so that the package
manager can properly receive content from the repositories of the
self-deployed Candlepin. This should have been done in the past, and it
will be critical for switching to SCA for testing: in SCA mode, all the
content of the organization is available by default, and some
repositories are also automatically enabled.

This means:
- installing the GPG key for RPM packages
- setting the baseurl in rhsm.conf