fix: adapt to the new state of insights-client.service on failure
There were recent changes to the insights-client.service systemd service
of insights-client:
https://github.com/RedHatInsights/insights-client/pull/240
This in practice means that insights-client.service now automatically
restarts itself in case of failure. Because of that, that the status of
the service on failure now is not "failed" but "activating", since it
will enabled again the next time insights-client.timer is triggered;
the effect in the cockpit plugin is that the detection for a failed
upload now does not work.

To adapt to the new situation, tweak the detection a bit: in addition to
the old logic (left there to support both old and new versions of
insights-client), a new logic is added to check that the systemd service
is "starting" (the cockpit mapping to the "activating" status) and its
last result is different than "success" (hence it failed somehow).

Also adapt the test a bit: the service needs explicit restarting now,
rather than a "simple" start.

Card ID: CCT-702