test: drop extra activation key code
The Python snippet to add a pool to an activation key in practice does
not work anymore for two reasons:
1) the test data in the self-deployed Candlepin changed
2) the test account used is based on SCA
and the result of the last POST in the script is:
  {
    "displayMessage" : "Multi-entitlement not supported for pool \"Pool [id: 8a8082e491b0fb160191b0fb37e20233, type: NORMAL, product: sfs, productName: Shared File System, quantity: 5]\"",
    "requestUuid" : "8e43664e-e720-4cb8-8a99-bfe3cd49af2c"
  }

Especially because of (2), it is possible to get rid of the script
altogether, and simply use the available activation keys.

There is no behaviour change.