Various fixes in state management
This improves the managing of state between webui and subscriptionsClient.
There is a new `addAlert` handler function which manages alerts and
fixes bug where an already present alert was rendered on each new render.

Also fixes createRoot() calling and the following error:
```
Warning: You are calling ReactDOMClient.createRoot() on a container that
has already been passed to createRoot() before. Instead,
call root.render() on the existing root instead if you want to update it.
```